+++ test-parts/09-api_frontend/sub/common-nodes.js
@@ -32,7 +32,6 @@ function testNodeStorageObject(storageObject)
 		currentProperty = storageObject[nodeManufacturer];
 		
 		testNodeObjectArrayStructure(currentProperty);
-		testNodeObjectArrayProperties(currentProperty);
 		testNodeObjectArrayContents(currentProperty);
 	}
 }