+++ changelog.md
@@ -1,3 +1,14 @@
 # Changelog
 
-All offline tests work after code merging.
+**./test-parts/08-api_requests/requests/req-admin_health.js**
+* Removed "File System Properties" test from 'handleFileSystem'
+
+---
+
+**./test-parts/08-api_requests/sub/common-storage.js**
+* Removed 'testDrivePropertiesArray' function.
+
+---
+
+**./test-parts/09-api_frontend/sub/common-nodes.js**
+* Removed 'testNodeObjectArrayProperties' call from 'testNodeStorageObject'